backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
openai
pydantic